+++ tests/testthat/test-text_nunit.R
@@ -35,14 +35,3 @@ test_that("text_ntoken can works on tokens", {
 test_that("text_ntoken handles NA and empty", {
     expect_equal(text_ntoken(c(NA, "")), c(NA, 0))
 })
-
-
-test_that("text_length includes NAs, text_ntoken doesn't", {
-    x <- c("A man, a plan.", "A \"canal\"?", "Panama!")
-    f <- text_filter(drop_punct = TRUE)
-    toks <- text_tokens(x, f)
-    ntok <- text_ntoken(x, f)
-    len <- text_length(x, f)
-    expect_equal(len, sapply(toks, length))
-    expect_equal(ntok, len - sapply(toks, function(t) sum(is.na(t))))
-})